+++ src/main/java/br/inatel/icc/idp/artefato/model/UserEntity.java
@@ -1,6 +1,5 @@
 package br.inatel.icc.idp.artefato.model;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -25,7 +24,7 @@ public class UserEntity {
     private final String email;
     private final String bio;
     private final Boolean isCrafter;
-    private final BigDecimal wallet;
+    private final Double wallet;
 
     @Relationship(type = "CRAFTED", direction = Direction.OUTGOING)
     private List<ProductEntity> product;
@@ -36,7 +35,7 @@ public class UserEntity {
     @Relationship(type = "FOLLOW", direction = Direction.OUTGOING)
     private List<UserEntity> follow;
 
-    public UserEntity(String name, String email, String bio, Boolean isCrafter, BigDecimal wallet) {
+    public UserEntity(String name, String email, String bio, Boolean isCrafter, Double wallet) {
         this.id = null;
         this.name = name;
         this.email = email;
